chore(origin-tls-client-auth): update generated types and methods

+++ src/resources/origin-tls-client-auth/origin-tls-client-auth.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../resource';
+import { isRequestOptions } from '../../core';
 import * as Core from '../../core';
 import * as HostnameCertificatesAPI from './hostname-certificates';
 import {
@@ -71,7 +72,7 @@ export class OriginTLSClientAuth extends APIResource {
     params: OriginTLSClientAuthCreateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<OriginTLSClientAuthCreateResponse> {
-    const { zone_id, ...body } = params;
+    const { zone_id = this._client.zoneId, ...body } = params;
     return (
       this._client.post(`/zones/${zone_id}/origin_tls_client_auth`, { body, ...options }) as Core.APIPromise<{
         result: OriginTLSClientAuthCreateResponse;
@@ -86,10 +87,20 @@ export class OriginTLSClientAuth extends APIResource {
    * @deprecated Use zone_certificates.list for zone-level certificates. This method will be removed in a future major version.
    */
   list(
-    params: OriginTLSClientAuthListParams,
+    params?: OriginTLSClientAuthListParams,
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<OriginTLSClientAuthListResponsesSinglePage, OriginTLSClientAuthListResponse>;
+  list(
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<OriginTLSClientAuthListResponsesSinglePage, OriginTLSClientAuthListResponse>;
+  list(
+    params: OriginTLSClientAuthListParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<OriginTLSClientAuthListResponsesSinglePage, OriginTLSClientAuthListResponse> {
-    const { zone_id } = params;
+    if (isRequestOptions(params)) {
+      return this.list({}, params);
+    }
+    const { zone_id = this._client.zoneId } = params;
     return this._client.getAPIList(
       `/zones/${zone_id}/origin_tls_client_auth`,
       OriginTLSClientAuthListResponsesSinglePage,
@@ -104,10 +115,22 @@ export class OriginTLSClientAuth extends APIResource {
    */
   delete(
     certificateId: string,
-    params: OriginTLSClientAuthDeleteParams,
+    params?: OriginTLSClientAuthDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<OriginTLSClientAuthDeleteResponse>;
+  delete(
+    certificateId: string,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<OriginTLSClientAuthDeleteResponse>;
+  delete(
+    certificateId: string,
+    params: OriginTLSClientAuthDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<OriginTLSClientAuthDeleteResponse> {
-    const { zone_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(certificateId, {}, params);
+    }
+    const { zone_id = this._client.zoneId } = params;
     return (
       this._client.delete(
         `/zones/${zone_id}/origin_tls_client_auth/${certificateId}`,
@@ -124,10 +147,19 @@ export class OriginTLSClientAuth extends APIResource {
    */
   get(
     certificateId: string,
-    params: OriginTLSClientAuthGetParams,
+    params?: OriginTLSClientAuthGetParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<OriginTLSClientAuthGetResponse>;
+  get(certificateId: string, options?: Core.RequestOptions): Core.APIPromise<OriginTLSClientAuthGetResponse>;
+  get(
+    certificateId: string,
+    params: OriginTLSClientAuthGetParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<OriginTLSClientAuthGetResponse> {
-    const { zone_id } = params;
+    if (isRequestOptions(params)) {
+      return this.get(certificateId, {}, params);
+    }
+    const { zone_id = this._client.zoneId } = params;
     return (
       this._client.get(
         `/zones/${zone_id}/origin_tls_client_auth/${certificateId}`,
@@ -231,7 +263,7 @@ export interface OriginTLSClientAuthCreateParams {
   /**
    * Path param: Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Body param: The zone's leaf certificate.
@@ -248,21 +280,21 @@ export interface OriginTLSClientAuthListParams {
   /**
    * Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 }
 
 export interface OriginTLSClientAuthDeleteParams {
   /**
    * Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 }
 
 export interface OriginTLSClientAuthGetParams {
   /**
    * Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 }
 
 OriginTLSClientAuth.OriginTLSClientAuthListResponsesSinglePage = OriginTLSClientAuthListResponsesSinglePage;
